Move l_filter to art.preprocessing
