+++ art/preprocessing/__init__.py
@@ -6,3 +6,5 @@ from art.preprocessing.standardisation_mean_std.standardisation_mean_std_pytorch
 from art.preprocessing.standardisation_mean_std.standardisation_mean_std_tensorflow import (
     StandardisationMeanStdTensorFlowV2,
 )
+from art.preprocessing.l_filter.l_filter import LFilter
+from art.preprocessing.l_filter.l_filter_pytorch import LFilterPyTorch
