+++ tests/preprocessing/test_l_filter.py
@@ -22,7 +22,7 @@ import logging
 import numpy as np
 import pytest
 
-from art.defences.preprocessor import LFilter
+from art.preprocessing import LFilter
 from tests.utils import ARTTestException
 
 logger = logging.getLogger(__name__)
